+++ storage-agent/src/BackupMesh.Storage.Service/Program.cs
@@ -56,6 +56,8 @@ builder.Services.Configure<BackupCommandOptions>(builder.Configuration.GetSectio
 builder.Services.Configure<AutomationSettingsOptions>(builder.Configuration.GetSection("AutomationSettings"));
 builder.Services.Configure<PairingOptions>(builder.Configuration.GetSection("Pairing"));
 builder.Services.Configure<PairingCertificateOptions>(builder.Configuration.GetSection("PairingCertificate"));
+builder.Services.Configure<IssuedCertificateOptions>(builder.Configuration.GetSection("IssuedCertificate"));
+builder.Services.Configure<SourceDisplayNameOptions>(builder.Configuration.GetSection("SourceDisplayName"));
 builder.Services.AddSingleton(mutualTls);
 builder.Services.AddSingleton(sp => sp.GetRequiredService<Microsoft.Extensions.Options.IOptions<StorageOptions>>().Value);
 builder.Services.AddSingleton(sp => sp.GetRequiredService<Microsoft.Extensions.Options.IOptions<RestServerOptions>>().Value);
@@ -68,6 +70,8 @@ builder.Services.AddSingleton(sp => sp.GetRequiredService<Microsoft.Extensions.O
 builder.Services.AddSingleton(sp => sp.GetRequiredService<Microsoft.Extensions.Options.IOptions<AutomationSettingsOptions>>().Value);
 builder.Services.AddSingleton(sp => sp.GetRequiredService<Microsoft.Extensions.Options.IOptions<PairingOptions>>().Value);
 builder.Services.AddSingleton(sp => sp.GetRequiredService<Microsoft.Extensions.Options.IOptions<PairingCertificateOptions>>().Value);
+builder.Services.AddSingleton(sp => sp.GetRequiredService<Microsoft.Extensions.Options.IOptions<IssuedCertificateOptions>>().Value);
+builder.Services.AddSingleton(sp => sp.GetRequiredService<Microsoft.Extensions.Options.IOptions<SourceDisplayNameOptions>>().Value);
 builder.Services.AddSingleton<StorageStateMachine>();
 builder.Services.AddSingleton<BackupJobStore>();
 builder.Services.AddSingleton<BackupCommandQueue>();
@@ -77,6 +81,8 @@ builder.Services.AddSingleton<RevokedSourceStore>();
 builder.Services.AddSingleton<PairingSessionStore>();
 builder.Services.AddSingleton<PairingAttemptThrottle>();
 builder.Services.AddSingleton(pairingCertificateAuthority);
+builder.Services.AddSingleton<IssuedCertificateStore>();
+builder.Services.AddSingleton<SourceDisplayNameStore>();
 builder.Services.AddSingleton<SourceCatalogStore>();
 builder.Services.AddSingleton<StorageConfigurationStore>();
 builder.Services.AddSingleton<StoragePresenceStore>();